The LONG-awaited update. - Migrated all int-32 to use Longs (since most requests are unsigned ints, might as well future-proof it); - Added Minecraft request category; - Added game versions V2 request; - Rework requests and their caching;

+++ src/main/java/org/zeith/cfcore4j/categories/GetCategoriesRequest.java
@@ -6,32 +6,32 @@ import org.zeith.cfcore4j.query.QueryBuilder;
 public class GetCategoriesRequest
 		implements IQueryContainer
 {
-	private Integer gameId;
-	private Integer classId;
+	private Long gameId;
+	private Long classId;
 
 	public static GetCategoriesRequest create()
 	{
 		return new GetCategoriesRequest();
 	}
 
-	public GetCategoriesRequest gameId(int gameId)
+	public GetCategoriesRequest gameId(long gameId)
 	{
 		this.gameId = gameId;
 		return this;
 	}
 
-	public GetCategoriesRequest classId(int classId)
+	public GetCategoriesRequest classId(long classId)
 	{
 		this.classId = classId;
 		return this;
 	}
 
-	public Integer gameId()
+	public Long gameId()
 	{
 		return gameId;
 	}
 
-	public Integer classId()
+	public Long classId()
 	{
 		return classId;
 	}
